backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.7.2
httpx==0.27.2
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
groq==0.11.0
tiktoken==0.7.0
cohere==5.9.2
fpdf2==2.8.1
pypdf
python-docx
python-pptx
